+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to `simple-excel` will be documented in this file
 
 ## 1.2.0 - 2019-11-29
 
-- add `streamToBrowser`
+- add `streamDownload` and `toBrowser`
 
 ## 1.1.0 - 2019-10-27